partikel/requirements.txt
attrs==23.1.0
click==8.1.3
dm-tree==0.1.8
matplotlib==3.7.1
numba==0.57.1
pyro-ppl==1.8.6
pytest==7.4.0
scipy==1.11.1
torch==2.0.0
torch-scatter==2.1.2
